+++ docs/changelog.md
@@ -5,6 +5,7 @@
 - Do not mutate the input dataset when decoding ({pull}`226`)
 - Properly decode ellipsoids for the CF convention ({pull}`228`)
 - Rewrite the map widgets using `anywidget` ({pull}`248`)
+- Resolve ellipsoid names on user input ({pull}`257`)
 
 ## 0.6.0 (2026-02-05)